+++ src/lib/components/Gallery.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { onMount } from "svelte";
+  import { onDestroy, onMount } from "svelte";
   import { urlForImage } from "$lib/config/sanity";
 
   import BiggerPicture from "bigger-picture/svelte";
@@ -11,76 +11,69 @@
   let galleryImages = images.map((image: any) => {
     if (!image.vertical) {
       return {
-        src: urlForImage(image).height(750).width(1000).quality(90).url(),
+        img: urlForImage(image).height(1333).width(2000).quality(80).url(),
         thumb: urlForImage(image).width(400).quality(50).url(),
         alt: image.alt,
-        height: 750,
-        width: 1000,
+        height: 2000,
+        width: 3000,
         photographer: image.photographer?.name,
         photographerUrl: image.photographer?.imageUrl,
       };
     } else if (image.vertical) {
       return {
-        src: urlForImage(image).height(1000).width(750).quality(90).url(),
+        img: urlForImage(image).height(2000).width(1333).quality(80).url(),
         thumb: urlForImage(image).width(400).quality(50).url(),
         alt: image.alt,
-        height: 1000,
-        width: 750,
+        height: 2000,
+        width: 1333,
         photographer: image.photographer?.name,
         photographerUrl: image.photographer?.imageUrl,
       };
     }
   });
 
+  let bp: any;
+
   onMount(() => {
-    // initialize
-    let bp = BiggerPicture({
-      target: document.body,
+    // initialize BiggerPicture
+    bp = BiggerPicture({
+      target: document.getElementById("inline")!
     });
 
-    // grab image links
-    let imageLinks = document.querySelectorAll(`a.${id}`);
-    // add click listener to open BiggerPicture
-    for (let link of imageLinks) {
-      link.addEventListener("click", openGallery);
-    }
+    // open inline gallery
+    bp.open({
+      items: galleryImages,
+      inline: true,
+      scale: 1,
+      intro: "fadeup",
+      noClose: true,
+    });
+  });
 
-    // open BiggerPicture
-    function openGallery(e: Event) {
-      e.preventDefault();
-      bp.open({
-        items: imageLinks,
-        el: e.currentTarget!,
-      });
+  onDestroy(() => {
+    if (bp) {
+      bp.close();
     }
   });
 </script>
 
-<section class="w-full mx-auto">
-  <div
-    id="images"
-    class="columns-2 sm:columns-3 md:columns-4 lg:columns-4 xl:columns-5"
-  >
-    {#each galleryImages as image}
-      <a
-        class={id}
-        href={image?.src}
-        data-img={image?.src}
-        data-thumb={image?.thumb}
-        data-height={image?.height}
-        data-width={image?.width}
-        data-alt={image?.alt}
-        data-caption={`Fotografía de ${image?.photographer} (<a class="underline" target="_blank" href=${image?.photographerUrl}>Instagram</a>)`}
-        aria-label={image?.alt}
-      >
-        <img
-          class="mb-2 rounded-sm"
-          loading="lazy"
-          width="400"
-          src={image?.thumb}
-          alt={image?.alt}
-        />
-      </a>
-    {/each}
-  </div>
+<section class="border-2 border-primary z-10">
+  <div id="inline" class="inline-gallery" />
 </section>
+
+<style>
+  * {
+    box-sizing: border-box;
+  }
+
+  .inline-gallery {
+    position: relative;
+    background: #222;
+    overflow: hidden;
+  }
+  .inline-gallery:before {
+    content: "";
+    padding-bottom: 66.66%;
+    display: block;
+  }
+</style>